# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OrderNetwork = typeof OrderNetwork[keyof typeof OrderNetwork];


export const OrderNetwork = {
  MTN: 'MTN',
  Vodacom: 'Vodacom',
  Telkom: 'Telkom',
  Cell_C: 'Cell C',
} as const;

export type OrderVoucherType = typeof OrderVoucherType[keyof typeof OrderVoucherType];


export const OrderVoucherType = {
  Blu: 'Blu',
  OTT: 'OTT',
  '1Voucher': '1Voucher',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  Pending: 'Pending',
  Processing: 'Processing',
  Completed: 'Completed',
} as const;

export interface Order {
  id: number;
  phone: string;
  network: OrderNetwork;
  voucherType: OrderVoucherType;
  voucherPin: string;
  whatsapp: string;
  status: OrderStatus;
  createdAt: string;
}

export type OrderInputNetwork = typeof OrderInputNetwork[keyof typeof OrderInputNetwork];


export const OrderInputNetwork = {
  MTN: 'MTN',
  Vodacom: 'Vodacom',
  Telkom: 'Telkom',
  Cell_C: 'Cell C',
} as const;

export type OrderInputVoucherType = typeof OrderInputVoucherType[keyof typeof OrderInputVoucherType];


export const OrderInputVoucherType = {
  Blu: 'Blu',
  OTT: 'OTT',
  '1Voucher': '1Voucher',
} as const;

export interface OrderInput {
  /**
     * @minLength 10
     * @maxLength 15
     */
  phone: string;
  network: OrderInputNetwork;
  voucherType: OrderInputVoucherType;
  /**
     * @minLength 4
     * @maxLength 20
     */
  voucherPin: string;
  /**
     * @minLength 10
     * @maxLength 15
     */
  whatsapp: string;
}

export type OrderTrackingNetwork = typeof OrderTrackingNetwork[keyof typeof OrderTrackingNetwork];


export const OrderTrackingNetwork = {
  MTN: 'MTN',
  Vodacom: 'Vodacom',
  Telkom: 'Telkom',
  Cell_C: 'Cell C',
} as const;

export type OrderTrackingVoucherType = typeof OrderTrackingVoucherType[keyof typeof OrderTrackingVoucherType];


export const OrderTrackingVoucherType = {
  Blu: 'Blu',
  OTT: 'OTT',
  '1Voucher': '1Voucher',
} as const;

export type OrderTrackingStatus = typeof OrderTrackingStatus[keyof typeof OrderTrackingStatus];


export const OrderTrackingStatus = {
  Pending: 'Pending',
  Processing: 'Processing',
  Completed: 'Completed',
} as const;

export interface OrderTracking {
  id: number;
  network: OrderTrackingNetwork;
  voucherType: OrderTrackingVoucherType;
  status: OrderTrackingStatus;
  createdAt: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  Pending: 'Pending',
  Processing: 'Processing',
  Completed: 'Completed',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  processingOrders: number;
  completedOrders: number;
  totalRevenue: number;
}

export interface Bundle {
  id: number;
  name: string;
  /** @minimum 0 */
  price: number;
  dataAmount: string;
  description: string;
  active: boolean;
}

export interface BundleInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 0 */
  price: number;
  /** @minLength 1 */
  dataAmount: string;
  description: string;
  active?: boolean;
}

export interface BundleUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  price?: number;
  /** @minLength 1 */
  dataAmount?: string;
  description?: string;
  active?: boolean;
}

export type ListAdminOrdersParams = {
search?: string;
};