# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Customer order operations
  - name: admin
    description: Admin-only operations
  - name: bundles
    description: Data bundle catalog
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /bundles:
    get:
      operationId: listBundles
      tags: [bundles]
      summary: List active data bundles
      responses:
        "200":
          description: List of bundles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bundle"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit a new voucher order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order's status by id (customer tracking, redacted of sensitive fields)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"
        "404":
          description: Order not found

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Log in to the admin panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid password

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Log out of the admin panel
      responses:
        "204":
          description: Logged out

  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check current admin session
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin), optionally filtered by search text
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Not authenticated

  /admin/orders/{id}:
    patch:
      operationId: updateAdminOrder
      tags: [admin]
      summary: Update an order's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Not authenticated
        "404":
          description: Order not found
    delete:
      operationId: deleteAdminOrder
      tags: [admin]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "404":
          description: Order not found

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Not authenticated

  /admin/bundles:
    get:
      operationId: listAdminBundles
      tags: [admin]
      summary: List all data bundles, including inactive ones (admin)
      responses:
        "200":
          description: List of bundles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bundle"
        "401":
          description: Not authenticated
    post:
      operationId: createBundle
      tags: [admin]
      summary: Create a data bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleInput"
      responses:
        "201":
          description: Bundle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        "401":
          description: Not authenticated

  /admin/bundles/{id}:
    patch:
      operationId: updateBundle
      tags: [admin]
      summary: Update a data bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleUpdate"
      responses:
        "200":
          description: Updated bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        "401":
          description: Not authenticated
        "404":
          description: Bundle not found
    delete:
      operationId: deleteBundle
      tags: [admin]
      summary: Delete a data bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
        "404":
          description: Bundle not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Order:
      type: object
      required:
        - id
        - phone
        - network
        - voucherType
        - voucherPin
        - whatsapp
        - status
        - createdAt
      properties:
        id:
          type: integer
        phone:
          type: string
        network:
          type: string
          enum: [MTN, Vodacom, Telkom, "Cell C"]
        voucherType:
          type: string
          enum: [Blu, OTT, 1Voucher]
        voucherPin:
          type: string
        whatsapp:
          type: string
        status:
          type: string
          enum: [Pending, Processing, Completed]
        createdAt:
          type: string
          format: date-time

    OrderInput:
      type: object
      required:
        - phone
        - network
        - voucherType
        - voucherPin
        - whatsapp
      properties:
        phone:
          type: string
          minLength: 10
          maxLength: 15
        network:
          type: string
          enum: [MTN, Vodacom, Telkom, "Cell C"]
        voucherType:
          type: string
          enum: [Blu, OTT, 1Voucher]
        voucherPin:
          type: string
          minLength: 4
          maxLength: 20
        whatsapp:
          type: string
          minLength: 10
          maxLength: 15

    OrderTracking:
      type: object
      required:
        - id
        - network
        - voucherType
        - status
        - createdAt
      properties:
        id:
          type: integer
        network:
          type: string
          enum: [MTN, Vodacom, Telkom, "Cell C"]
        voucherType:
          type: string
          enum: [Blu, OTT, 1Voucher]
        status:
          type: string
          enum: [Pending, Processing, Completed]
        createdAt:
          type: string
          format: date-time

    OrderStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [Pending, Processing, Completed]

    AdminCredentials:
      type: object
      required:
        - password
      properties:
        password:
          type: string

    AdminSession:
      type: object
      required:
        - authenticated
      properties:
        authenticated:
          type: boolean

    AdminStats:
      type: object
      required:
        - totalOrders
        - pendingOrders
        - processingOrders
        - completedOrders
        - totalRevenue
      properties:
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        processingOrders:
          type: integer
        completedOrders:
          type: integer
        totalRevenue:
          type: number

    Bundle:
      type: object
      required:
        - id
        - name
        - price
        - dataAmount
        - description
        - active
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
          minimum: 0
        dataAmount:
          type: string
        description:
          type: string
        active:
          type: boolean

    BundleInput:
      type: object
      required:
        - name
        - price
        - dataAmount
        - description
      properties:
        name:
          type: string
          minLength: 1
        price:
          type: integer
          minimum: 0
        dataAmount:
          type: string
          minLength: 1
        description:
          type: string
        active:
          type: boolean

    BundleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        price:
          type: integer
          minimum: 0
        dataAmount:
          type: string
          minLength: 1
        description:
          type: string
        active:
          type: boolean