# lib/api-zod/src/generated/types/bundleUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BundleUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minimum 0 */
  price?: number;
  /** @minLength 1 */
  dataAmount?: string;
  description?: string;
  active?: boolean;
}