# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderNetwork } from './orderNetwork';
import type { OrderStatus } from './orderStatus';
import type { OrderVoucherType } from './orderVoucherType';

export interface Order {
  id: number;
  phone: string;
  network: OrderNetwork;
  voucherType: OrderVoucherType;
  voucherPin: string;
  whatsapp: string;
  status: OrderStatus;
  createdAt: Date;
}