# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputNetwork } from './orderInputNetwork';
import type { OrderInputVoucherType } from './orderInputVoucherType';

export interface OrderInput {
  /**
     * @minLength 10
     * @maxLength 15
     */
  phone: string;
  network: OrderInputNetwork;
  voucherType: OrderInputVoucherType;
  /**
     * @minLength 4
     * @maxLength 20
     */
  voucherPin: string;
  /**
     * @minLength 10
     * @maxLength 15
     */
  whatsapp: string;
}